+++ examples/sharded_array_write_read.rs
@@ -1,5 +1,8 @@
 use itertools::Itertools;
-use zarrs::storage::ListableStorageTraits;
+use zarrs::storage::{
+    storage_transformer::{StorageTransformerExtension, UsageLogStorageTransformer},
+    ReadableWritableListableStorage,
+};
 
 fn sharded_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     use zarrs::{
@@ -17,21 +20,24 @@ fn sharded_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
 
     // Create a store
     // let path = tempfile::TempDir::new()?;
-    // let store = Arc::new(store::FilesystemStore::new(path.path())?);
-    // let store = Arc::new(store::FilesystemStore::new("tests/data/sharded_array_write_read.zarr")?);
-    let store = Arc::new(store::MemoryStore::new());
-
-    // let log_writer = Arc::new(std::sync::Mutex::new(
-    //     // std::io::BufWriter::new(
-    //     std::io::stdout(),
-    //     //    )
-    // ));
-    // let usage_log = Arc::new(UsageLogStorageTransformer::new(log_writer, || {
-    //     chrono::Utc::now().format("[%T%.3f] ").to_string()
-    // }));
-    // let store = usage_log
-    //     .clone()
-    //     .create_readable_writable_transformer(store);
+    // let mut store: ReadableWritableListableStorage = Arc::new(store::FilesystemStore::new(path.path())?);
+    // let mut store: ReadableWritableListableStorage = Arc::new(store::FilesystemStore::new("tests/data/sharded_array_write_read.zarr")?);
+    let mut store: ReadableWritableListableStorage = Arc::new(store::MemoryStore::new());
+    if let Some(arg1) = std::env::args().collect::<Vec<_>>().get(1) {
+        if arg1 == "--usage-log" {
+            let log_writer = Arc::new(std::sync::Mutex::new(
+                // std::io::BufWriter::new(
+                std::io::stdout(),
+                //    )
+            ));
+            let usage_log = Arc::new(UsageLogStorageTransformer::new(log_writer, || {
+                chrono::Utc::now().format("[%T%.3f] ").to_string()
+            }));
+            store = usage_log
+                .clone()
+                .create_readable_writable_listable_transformer(store);
+        }
+    }
 
     // Create a group
     let group_path = "/group";