+++ examples/zip_array_write_read.rs
@@ -8,13 +8,17 @@ use std::{
 use zarrs::{
     array::{Array, ZARR_NAN_F32},
     array_subset::ArraySubset,
-    storage::{ReadableStorageTraits, ReadableWritableStorageTraits, StoreKey},
+    storage::{
+        storage_transformer::{StorageTransformerExtension, UsageLogStorageTransformer},
+        ReadableStorageTraits, ReadableWritableListableStorage, ReadableWritableStorageTraits,
+        StoreKey,
+    },
 };
 
 // const ARRAY_PATH: &'static str = "/array";
 const ARRAY_PATH: &str = "/";
 
-fn write_array_to_storage<TStorage: ReadableWritableStorageTraits + 'static>(
+fn write_array_to_storage<TStorage: ReadableWritableStorageTraits + ?Sized + 'static>(
     storage: Arc<TStorage>,
 ) -> Result<Array<TStorage>, Box<dyn std::error::Error>> {
     use zarrs::array::{chunk_grid::ChunkGridTraits, codec, DataType, FillValue};
@@ -150,7 +154,25 @@ fn zip_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     let path = tempfile::TempDir::new()?;
     let mut zarr_dir = path.path().to_path_buf();
     zarr_dir.push("hierarchy.zarr");
-    let store = Arc::new(store::FilesystemStore::new(&zarr_dir)?);
+    let mut store: ReadableWritableListableStorage =
+        Arc::new(store::FilesystemStore::new(&zarr_dir)?);
+    if let Some(arg1) = std::env::args().collect::<Vec<_>>().get(1) {
+        if arg1 == "--usage-log" {
+            let log_writer = Arc::new(std::sync::Mutex::new(
+                // std::io::BufWriter::new(
+                std::io::stdout(),
+                //    )
+            ));
+            let usage_log = Arc::new(UsageLogStorageTransformer::new(log_writer, || {
+                chrono::Utc::now().format("[%T%.3f] ").to_string()
+            }));
+            store = usage_log
+                .clone()
+                .create_readable_writable_listable_transformer(store);
+        }
+    }
+
+    // Write the array to the store
     write_array_to_storage(store.clone())?;
 
     // Write the store to zip